+++ scripts/dotnet/OfflineFireRedAsrModelConfig.cs
@@ -1,0 +1,21 @@
+/// Copyright (c)  2025  Xiaomi Corporation (authors: Fangjun Kuang)
+
+using System.Runtime.InteropServices;
+
+namespace SherpaOnnx
+{
+    [StructLayout(LayoutKind.Sequential)]
+    public struct OfflineFireRedAsrModelConfig
+    {
+        public OfflineFireRedAsrModelConfig()
+        {
+            Encoder = "";
+            Decoder = "";
+        }
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Encoder;
+
+        [MarshalAs(UnmanagedType.LPStr)]
+        public string Decoder;
+    }
+}